Rename text_template_toolbar folder to textTemplateToolbar
