+++ app/packs/src/components/textTemplateToolbar/TextTemplateToolbar.js
